acpi: acpitables: remove irrelevant test case (LP: #1287891)
Although the ACPI specification in section 5.2.9 states that
the 32 bit and the 64 bit firmware controls should not be both
defined at the same time, the reality is that it is on a lot of
firmware and the the kernel uses the 64 bit version if both are
defined. So just remove this test as it causes more harm than
good.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>