fwts-test: arg-log-format-0001: make test less pedandic on line numbers
Line numbers can change depending on arch and build specific information
dumped out at the start of the test.  So make this test less pedantic on
line number checking.  We are just interested in the format and not
the actual content of some fields.  This fixes build test failures on
ARM.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>