fwts-test: arg-show-progress-dialog: remove version from test
The version test varies in the number of tests being run on different
architectures which can cause a false positive failure on ppc64el.
So remove the version test and just run the oops and klog test in this
test which in theory are the same across all achitectures.

Also remove the results.log which was accidentally introduced and
we don't need.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>