acpica: ppc64el needs to be built as a 64 bit architecture
ppc64el was segfaulting because the build was defaulting to
a 32 bit build.  This is a temporary fix for the current
release of fwts, I hope to get this fixed upstream in ACPICA
some time soon.  With this fix the syntaxcheck and disassemble
regression tests were failing because iasl was segfaulting.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>