acpi: acpidump: fix dump offsets in various table dumps (LP: #1293613)
fwts acpidump is reporting incorrect table offset addresses
for several tables: BERT, CPEP, BGRT, GTDT, RASF and PCCT. I
unfortunately kept on repeating the same bug as I introduced
new tables. The index offset for these should always be zero.

When testing this on ARM I also hit a Bus Error because of an
unaligned 64 bit access when fetching data from some of these
tables, so memcpy the data to the uin64_t target because the
source 64 bit value is possibly not naturally aligned on some
platforms

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>