acpi: acpitables: more ACPI table header sanity checking
Some recent firmware contained corrupt XSDT that pointed to garbage
and fwts was not catching this.  Extend the acpitables test to do
some simple sanity checks to see if the table header is a suitable
size and also if some of the human readable fields contain garbage.

The ACPI specification seems vague about the vendor specific fields
that we check here, but I think if they don't contain valid 7 bit
ASCII text then they may be corrupt or the RSDT/XSDT may be pointing
to garbage, so it is worth reporting any weird looking fields.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>