efi_runtime: Do not pass user addresses to firmware
Currently there's some inconsistency with how arguments are passed to
the firmware from the efi_runtime driver. Some values have the standard
get_user()/put_user() calls, others do not.

Passing userspace pointers directly to the firmware is a bug because
those addresses may fault. And if they are going to fault we'd like to
know about it in the kernel rather than at some later time when
executing in firmware context.

Furthermore, beginning with v3.14 of the kernel the current tests
actually cause the kernel to crash because firmware calls are now done
with their own, entirely separate, page tables - no user addresses are
mapped during execution of runtime services.

This change doesn't require predication on a particular kernel version
because the efi_runtime should really have always done this copying
to/from userspace for every argument of the runtime services.

This patch is heavily based on one from Borislav.

Cc: Borislav Petkov <bp@alien8.de>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Borislav Petkov <bp@suse.de>
Acked-by: Ivan Hu <ivan.hu@canonical.com>