fwts-test: arg-show-progress-dialog-0001: fix armhf failures
The progress stats have more precision and show the length of
test duration.  This however breaks the testing on armhf because
of different floating point % progress values and also the longer
test runs on armhf.  To accommodate these differences, I've added
a couple of sed filters to normalise the % progress and test
duration output so that this test will work on all architectures.

Tested on armhf and amd64.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>