Add support for different power methods to suspend
pm-utils is now deprecated and, as a result, we should rely on
either Logind (where available) or on sysfs.

While autodetection defaults to either Logind or sysfs (preferring
The former), it is also possible to pass the --pm-method parameter
to specify one of the following methods:

logind, sysfs, pm-utils

Note: quirks are only available when using pm-utils

This makes a dependency on GLib necessary (only to avoid using
the C Dbus bindings).

Similar code should be written to handle S4.

The new functions in fwts_pipeio.c and fwts_stringextras.c come
from Logind's source code.
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>