lib: fwts: add more sanity checks when fetching the RSDP (LP: #1372849)
The --rsdp option allows a user to specify the RSDP address, which
could be erroneous. Also, it is possible (but very unlikely) that
the RSDP in set incorrectly at boot time. Currently fwts_acpi_get_rsdp()
does some checking but fails to return NULL on failure so fwts segfaults.
Also fwts assumes that the RSDP address is readable, but we should also
catch any segv faults in case it is not.

This patch adds some "safe" memory helpers that can catch segfaults and
uses these in the table loading mmap'd reads and copies.  The patch also
ensures NULL is returned if the RSDP is fails any sanity checks.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>