bios: mpcheck: reduce error level if ACPI is being used (LP: #1373340)
http://en.wikipedia.org/wiki/MultiProcessor_Specification states:

"Since most newer machines support Advanced Configuration and Power
Interface (ACPI) which subsumes the MPS functionality, MPS has for
the most part been supplanted by ACPI. MPS can still be useful on
machines or with operating systems that do not support ACPI."

fwts should therefore indicate that any errors in the mptables on an
ACPI compliant system should be LOW (ignored) if the ACPI LAPIC or
IOAPIC configurations are being used.

The best we can do for now is to check this by searching the kernel log
to see if ACPI is being used and providing more useful information
on the mpcheck if MP tables are not being used.  If this is so, we
also lower the error level to LOW for any MP table issues since they
are not critical on an ACPI based system.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>