acpi: acpidump: don't assume FPDT tables are correct about their record sizes (LP: #1375258)
Section 5.2.23.4 and 5.2.23.5 Performance Record Formats of the
ACPI specification states that the length field is a minimum
of 16 bytes for these two record types. Currently fwts believes
this data from the firmware, however, some firmware has been known
to set this to zero causing fwts to loop on the dumping of multiple
records.

fwts should instead not trust this field and anything less than 16
bytes should be dumped out as a hex dump since it can't be decoded.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>