fwts: catch segfaults and bus errors and dump stack
Bugs such as LP#1376448 have been hard to track down because
the target machine is not accessible and the user is not easily
able to pass over core dump information.  In an attempt to at
least identify where a fault is occurring, this patch adds in
a SIGSEGV and SIGBUS handler that will dump to stderr a stack
trace.

This patch also re-works some of signal() handlers in fwts to
ensure we save and restore the dump stack handler in a more
generic way for fwts. This also tidies up the old fashioned
signal() handler with the more modern signal handling mechanisms.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>