efi_runtime: ensure we don't allocate a zero byte buffer (LP: #1429890)]
When requesting a test on a zero sized buffer we must ensure that a
buffer is allocated and is of zero size when passed to the EFI service.

If we kalloc zero bytes, then a ZERO_SIZE_PTR is returned, which has
no meaning to the EFI service, so one strategy is to convert this to
a NULL before passing it to the EFI service. However, this is handled
by the service as an invalid parameter and returns with
EFI_INVALID_PARAMETER rather than EFI_BUFFER_TOO_SMALL (the latter
being what we are trying to actually exercise).

The best way forward is to allocate a buffer that is 1 wide char too
long, and return back the buffer offset by 1 wide char (i.e skipping
the leading 2 bytes). This allows us to always return a valid sized
buffer that is passed to the EFI service, even with zero length requests.
For zero sized buffers, if EFI writes into the zero sized buffer it will
corrupt the member following the buffer.  I believe further checks should
probably be added to also sanity check for these potential buffer overruns.

Across various platforms this arrangement trips the expected
EFI_BUFFER_TOO_SMALL error return, as originally expected in the fwts userspace
test case.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>