lib: fwts_alloc: provide fallback low memory allocator strategy (LP: #1452168)
The fwts low memory allocator made some assumptions that
/proc/self/maps was always available for parsing to gain
some smarts on how to do a 32 bit memory allocation.
Apparently some platforms this is not available or can't
be parsed or just does not provide enough free memory
slots so this fails.

Instead, we need to have a fall-back dumb low memory
allocator scheme that walks through memory trying to find
free regions using mincore() to do the probing.

Thanks to Suravee Suthikulpanit for finding, debugging and testing
this fix.

Tested-by: Suravee Suthikulpanit <suravee.suthikulpanit@amd.com>
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>