fwts-test: arg-help-0001: modify regex to match upper/lower case version
It seems that this test breaks if we have a lowercase version tag,
such as v15.03.01 that slipped into the previous release. While we should
probably not allow lower case version tags it does seem sensible to make
the test more flexible to also match lower case version tags just in
case this happens again in the future.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>