lib: fwts_acpi_tables: enforce stricter table loading checks
Coverty Scan is warning that malicious or malformed tables being read in
from file could cause problems with the low 32 bit allocator. Add some
stricter checking of the size of the buffer being read and also the total
ACPI table size.  ACPI stipulates that the table size is a 32 bit integer
(which is huge), but we need to keep to that maximum upper limit to be
in spec.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>