fwts: Use linux perf counters for cpu benchmarking
Currently, we use a loop counter to measure cpu performance. There are a
couple of drawbacks of this:

1) the numbers are fairly arbitrary, and don't reflect any external
values; and

2) the results are subject to peturbation from other tasks.

This change uses the linux perf counters to measure actual CPU cycles,
where available. This means we get numbers that are meaningful, and will
we'll get the performance value even if our process has been scheduled
off the CPU.

Now, the new results should match the actual cpu frequencies:

  CPU 0: 16 CPU frequency steps supported.
   Frequency | Relative Speed |  Cycles    | Bogo loops
  -----------+----------------+------------+-----------
   1.200 GHz |      36.5 %    | 1197262295 |    114643
   1.300 GHz |      39.5 %    | 1297048295 |    124214
   1.400 GHz |      42.6 %    | 1396813066 |    133889
   1.500 GHz |      45.6 %    | 1496592732 |    143456
   1.600 GHz |      48.6 %    | 1596393751 |    150124
   1.700 GHz |      51.7 %    | 1696137977 |    162350
   1.800 GHz |      54.7 %    | 1795918438 |    172071
   1.900 GHz |      57.8 %    | 1895685372 |    181249
   2.000 GHz |      60.8 %    | 1995453204 |    191176
   2.100 GHz |      63.8 %    | 2095225345 |    200753
   2.200 GHz |      66.9 %    | 2194993645 |    210282
   2.300 GHz |      69.9 %    | 2294780535 |    219945
   2.400 GHz |      73.0 %    | 2394544770 |    229664
   2.500 GHz |      76.0 %    | 2494315101 |    239121
   2.600 GHz |      79.0 %    | 2594055116 |    248536
   2.601 GHz |     100.0 %    | 3281891417 |    314546

  CPU 0 performance scaling OK

Because we still want benchmarks to work when perf counters are not
available, we still measure the loop count. This means we need to
represent test results as a struct, with cycles and loops members.

When tests need a scalar performance value,
fwts_cpu_benchmark_get_best_result will return the most accurate result
measured.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com
Acked-by: Alex Hung <alex.hung@canonical.com>