cpu/cpufreq: Don't reparse cpu information
Currently, we do a whole lot of scanning the /sys/devices/system/cpu
directory to parse cpu information, multiple times.

This change does a parse once on ->init, and populates a structure of
cpu info. The tests then query this structure instead of reparsing /sys.

We remove the unnecessary sched_setaffinity when we change a core's
cpu_frequency.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canaonical.com>