cpu/cpufreq: Only run performance tests on master threads
On a (single-socket) openpower machine, the cpufreq tests take nearly an
hour:

  $ time sudo ./src/fwts cpufreq
  Running 1 tests, results appended to results.log
  Test: CPU frequency scaling tests.
    CPU frequency performance tests.                        1 passed

  real	58m43.334s
  user	58m40.118s
  sys	0m0.366s

This is because we have 44 possible frequencies:

  $ wc -w scaling_available_frequencies
  44 scaling_available_frequencies

on 80 cpus:

  $ getconf _NPROCESSORS_ONLN
  80

That's a total of 3520 individual benchmarks, at one second per test.

However, those 80 cpus that linux reports are acually threads: 8 threads
per core, on a 10 core system. Because each of these threads is on the
same core, they share the same cpufreq control:

  $ ls -ld cpu{0..7}/cpufreq
  drwxr-xr-x 3 root root 0 May 21 03:06 cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu1/cpufreq -> ../cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu2/cpufreq -> ../cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu3/cpufreq -> ../cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu4/cpufreq -> ../cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu5/cpufreq -> ../cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu6/cpufreq -> ../cpu0/cpufreq
  lrwxrwxrwx 1 root root 0 May 21 01:29 cpu7/cpufreq -> ../cpu0/cpufreq

So, rather than running the exact same test 8 times over on each core,
this change only runs the cpufreq test on the 'master' threads.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>