fwts: Use gettimeofday for cpu benchmark
Currently, we use time() to measure out one second of time for a CPU
benchmark. This means we have to wait around one second for a new time
value before performing the test, and means we have to assume that we've
been running for about one second.

This change uses gettimeofday() instead; this way, we don't need to wait
before starting a test, and get millisecond resolution when calculating
the performance value.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>