lib: fwts_klog: fix vector size and handle errors from pcre_exec (LP: #1461520)
On an 32 bit x86 platform another stack smashing bug occurs where
pcre_exec scribbles over the stack because the vector being passed
to pcre_exec is not a multiple of 3 in size (as the API requires).

Make the vector overly large multiple of 3 to fix this. This is the same
fix as in an earlier bug (LP: #1401184), however, I failed to fix all
the occurrances of the original bug, hence we got bitten again by this.

Also, move the define for VECTOR_SIZE to the top of the source just to
be a bit tidier.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>