acpi: allow various ACPI table tests to run without root access
ACPI tables can be accessed without the need of the FWTS_FLAG_ROOT_PRIV
flag because the tables can be loaded from a file.  The root access check
is performed at table load time, so if the user is loading them from
the firmware or via /sys/firmware/acpi/tables then the fwts lib will
do the checks and the tests will be skipped if there are insufficient
access rights automatically.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>