lib: make acpidump parser more robust (LP: #1471202)
When given data that contains more than one space between each hex
byte the parser has a conflict between a fixed format input and less
than 16 bytes being scanned.  Since the data being allocated on
a short row is less than 16 bytes and the offsets always increment
by 16 bytes we get a mismatch between where the parser is writing
the data and the size of the allocated table.

Make the parser more robust:

1. Only parse rows that match the acpidump format exactly
2. Abort if offsets don't increment as we expect
3. Treat short rows of less than 16 bytes as end-of-table

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>