Add in the notion of ACPI compliance tests.
This patch adds in the --acpicompliance option.  The idea behind this
is to use existing ACPI tests, and add to them a number of more specific
tests that are designed to verify that the ACPI tables being used are in
compliance with the ACPI specification.

This is not the same as --acpitests.  In that case, we are generally looking
for issues in ACPI tables that we have seen in the past and we want to be
sure not to see them again.  In some cases, those are violations of the spec,
but they are not necessarily so.

There are no tests added by this patch, but just the framework needed to
be able to add them in subsequent patch sets.

Signed-off-by: Al Stone <al.stone@linaro.org>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>