lib: fwts_acpi_table: Fix FADT update in creation of fake FACS table.
If a FACS ACPI table is not found, a fake one is created and
a fake physical address is assigned to it. This fake physical
address is written to FIRMWARE_CTRL and X_FIRMWARE_CTRL fields
of FADT table. The problem is that the checksum of the FADT
table is not updated to reflect the change in FIRMWARE_CTRL
and X_FIRMWARE_CTRL fields. Thus, the 'checksum' test falsely
reports:

FAILED [MEDIUM] ACPITableChecksum: Test 1, Table FACP has
incorrect checksum, expected 0xe8, got 0xcf.

This patch changes the update of the FADT table. If one of
X_FIRMWARE_CTRL, FIRMWARE_CTRL fields is nonzero then that
value is used as physical address of FACS table. If both
fields are zero then a fake physical address is assigned
to FACS, written to FADT table and the checksum of the FADT
table is updated.

Signed-off-by: jvohanka <jvohanka@redhat.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>