lib: iasl: rework IASL interface to ACPICA
This set of changes re-works the iasl interface in fwts in preperation
for a re-working of acpidump to use IASL rather than hand-write our
own table dumping code.

The main objective of this change was to remove the awful original
design of referencing the tables to be disassembled by a table index
and instead allow fwts tests to reference the acpi table via the
info descriptor.  This allows more context to be passed around to
help simplify the next set of changes when we want to disassemble
not only DSTD and SSDT but any table type.

Unfortunately this change requires touching a few places from the
core ACPICA <--> fwts lib layer and the fwts lib <---> fwts test
layer.

One final change was to add a boolean flag to inform IASL to either
to use tables containing AML for exernal referencing or not. Only
disassembly of DSDT and SSDT should require this to be enabled.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>