acpi: acpidump: use IASL rather than fwts custom table dump code
Don't duplicate ACPI dumping and instead use IASL.  This reduces
the maintenance effort and the code size.  The output format has
been changed now that we are using IASL output, but now fwts is
in-sync with the core ACPICA standard output, so I think that is
a benefit rather than a problem.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>