s3 + klog: make log scanning more robust for S3 (LP: #1522292)
We sometimes get klog messages not appearing in the kernel log
when written by userspace via /dev/kmsg on recent kernels. I've
not yet figured out why that is, but a simple workaround for
the moment is to make the kernel log scanning for the FWTS S3
going to suspend marker is to check for other messages fwts
inserts into the log since at least one of these messages always
ends up in the log.

Also, add a kernel log priority level and fwts: prefix to all
messages from fwts that get written to /dev/ksmg for easier
identification.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>