ACPI: Add ASPT test
This table is not well described anywhere, however it is a frequently
used table on AMD machines and the format is relatively simple set of
4 32 bit addresses.  This table has been discussed on the ACPICA
devel mailing list:

https://lists.acpica.org/pipermail/devel/2015-November/000850.html

and this description matches the various acpi dumps of this table
on AMD machines that I have access too. I believe the table refers
to an AMD performance monitoring feature.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>