fwts: wakealarm: reset wakealarm timer at end of test (LP: #1534003)
The wakealarm test is setting the wakealarm timer but does not
restore it to the original state once the test is complete. This
fix saves the RTC wakealarm state and restores it again. I've added
a 5th test that does the restore and sanity checks this just to do
the job properly.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>