dmi: dmicheck: add some simple sanity checks to table size
The DMI table size should be sanity checked as an incorrect value
from the header may provide a bogus size and we could end up allocating
a buffer that will lead to a kernel OOM killer on fwts. The 32 bit
case the DMI tables are limited to 64K anyhow, but for the 64 bit case
I've limited this to 0xffffff rather than the 32 bit limit of 0xffffffff
as I really doubt tables are going to be larger than this.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>