lib: fwts_args: use safer strncat than strcat
Fix warning from static analysis that was complaining that strcat
is not safe, and use strncat instead.  Replace a strncat with
a few simply char ptr assignments since we know at that point
this is safe (we have enough buffer space) and we only have a few
chars to set.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>