Start defining FWTS architectures as variables
Add in fwts_arch.h and a typedef for fwts_architecture so that we can
start referring to the architectures used as variables.  Later on we
will define host and target architectures to make it easier for tests
to adapt their behavior.  What we want the tests to be able to do is
change a test -- if necessary -- based on the architecture supported
by the firmware being tested.  For example, if hardware reduced mode
is not being used on arm64, it is an error; on ia64 or x86, it may
not matter.

Signed-off-by: Al Stone <al.stone@linaro.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>