acpi: fadt: check 32/64 bit PM register addresses for inconsistency (LP: #1536606)
A HP BIOS was found in the wild that had a 64 bit PM register that was
different to the 32 bit PM register, and the 32 bit address was correct
but the 64 bit address was not. This is unreconcileable. Which one is
correct?

Upstream ACPICA commit 4da56eeae0749dfe8491285c1e1fad48f6efafd8 defaulted
to using 64 bit address by default which is now causing systems to become
unbootable (c.f. LP#1529381).

I think it is prudent to check for these inconsistencies as yet another
level of sanity checking.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>