lib: fwts_klog: handle the case where klog_old is empty list
scan-build detected that an empty klog_old list would lead to l_old_last
being null and hence we get a null pointer dereference when assigning
old to the data in the null list. Check for a empty list and ensure
l_new is initialized to NULL for this corner case.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>