acpica: fix disassembly on i386 platforms (LP: #1547469)
On i386 builds on Xenial I'm seeing:

*** Error in `/home/king/repos/fwts/src/.libs/lt-fwts': double free or
  corruption (!prev): 0x08bfb010 ***

this occurs because AcpiGbl_ParseOpRoot is being deleted and not being
set back to NULL, so subsequent parsing of externally loaded tables
append to this non-null tree with the last Op nodes and this causes
issues when we free these objects twice on the subsequent call of
AcpiPsDeleteParseTree.  We need to always ensure AcpiGbl_ParseOpRoot
is NULL after deleting the parse tree from it.

The second set of fixes syncs the disassembly code with the current
iASL main disassbly idoims.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>