FADT: expand compliance checks for DSDT and X_DSDT fields
Expand the testing of the DSDT address -- and by extension, the
X_DSDT address field -- to check for full compliance with the
spec.  Only one or the other may be used at any one time, per 6.1,
but we also have to acknowledge there are tables that do use both
the 32- and 64-bit values.  At that point, we re-use parts of the
existing test to verify that they are at least consistent.

Signed-off-by: Al Stone <al.stone@linaro.org>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>