FADT: add compliance tests for reserved fields, PM profile, reduced hardware
Add three new compliance tests that verify that the FADT reserved fields
are zero as they should be, that the PM (Power Management) profile is a
legal value, and that all of the fields and flags that are to be ignored
when the OSPM is in reduced hardware mode are zero.

Note that the spec says to ignore many FADT fields when in reduced hardware
mode.  As of 6.1, it does not explicitly state that they must be zero.  But,
zero seems to be a reasonable value -- or at least not destructive -- on the
OSPMs that use them when they should be ignored.  The case can also be made
that the fields should be zero; section 5.2.1.4 indicates ignored fields
should be treated the same as reserved fields, which _are_ required to be
zero.  As a practical matter, some vendors may have chosen to sidestep the
spec here and set reduced hardware mode but used the ignored field anyway.

Signed-off-by: Al Stone <al.stone@linaro.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>