v2 - Introduce olog scan, to check OPAL msglog.
This feature adds the olog option which will scan for patterns defined
in the olog.json data file and compare against the PPC OPAL firmware
msglog. Extending similar capabilities of the klog functionality,
this feature also allows customization to specify any file as the
source and any JSON data file as the patterns.

Signed-off-by: Deb McLemore <debmc@linux.vnet.ibm.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>