devicetree: add infrastructure for device-tree tests
This change introduces a little infrastructure for tests that consume
data from the device tree. This involves a few things:

An autoconf check for libfdt. If this is available, we define the
HAVE_LIBFDT makefile variable and config define.

A new member of fwts_framework to hold device tree data.

A new source (fwts_devicetree.{c,h}) in the fwts core code. At present,
this has one function to load & parse a device tree. We call this before
running tests, so that the device tree data is available to the tests.

Using the makefile variable, we conditionally compile tests that may
require libfdt. This way, we don't introduce a dependency for building
fwts for non-devicetree platforms.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>